+++ extensions/raycast-ollama/src/api/ollama.ts
@@ -1,84 +1,64 @@
 import {
-  OllamaApiTagsResponse,
   OllamaApiGenerateResponseUndone,
   OllamaApiGenerateResponseDone,
-  OllamaApiCreateRequestBody,
   OllamaApiGenerateRequestBody,
+  OllamaApiEmbeddingsResponse,
+  OllamaApiTagsResponse,
+  OllamaApiPullResponse,
 } from "./types";
 import {
   ErrorOllamaCustomModel,
   ErrorOllamaModelNotInstalled,
   ErrorOllamaNotInstalledOrRunning,
-  ErrorOllamaCantCreateCustomModel,
   MessageOllamaModelNotInstalled,
 } from "./errors";
-import { environment } from "@raycast/api";
 import fetch from "node-fetch";
 import { EventEmitter } from "stream";
 
 /**
- * Verify if a specific model is installed.
- * @param {string} model - Model name.
- * @returns {Promise<[boolean, string]>} [true, model] if model is installed, [false, model] if not.
+ * Get installed model.
+ * @returns {Promise<OllamaApiTagsResponse>} List of installed models.
  */
-async function OllamaApiTagsVerify(model: string): Promise<[boolean, string]> {
+export async function OllamaApiTags(): Promise<OllamaApiTagsResponse> {
   const url = "http://127.0.0.1:11434/api/tags";
-  const ModelMap = new Map<string, string>([
-    ["raycast_orca:3b", "orca:latest"],
-    ["raycast_llama2:7b", "llama2:latest"],
-    ["raycast_llama2:13b", "llama2:13b"],
-  ]);
 
-  const models = await fetch(url)
-    .then((response) => response.text())
+  const data = await fetch(url)
+    .then((response) => response.json())
     .then((output): OllamaApiTagsResponse => {
-      return JSON.parse(output);
+      return output as OllamaApiTagsResponse;
     })
     .catch((err) => {
       console.error(err);
       throw ErrorOllamaNotInstalledOrRunning;
     });
 
-  let modelIsInstalled = false;
-  models.models.forEach((row) => {
-    if (ModelMap.get(model) === row.name) {
-      modelIsInstalled = true;
-    }
-  });
-  return [modelIsInstalled, ModelMap.get(model) as string];
+  return data;
 }
 
 /**
- * Create a custom model from a file template.
- * @param {string} model - Model name. Can be one of the following: raycast_orca:3b, raycast_llama2:7b, raycast_llama2:13b.
- * @returns {Promise<boolean>} true if model is created, false if not.
+ * Delete model.
+ * @param {string} model Model name.
  */
-function OllamaApiCreateModel(model: string): Promise<boolean> {
-  const url = "http://127.0.0.1:11434/api/create";
-  const ModelFileMap = new Map<string, string>([
-    ["raycast_orca:3b", `${environment.assetsPath}/prompt/raycast_orca_3b`],
-    ["raycast_llama2:7b", `${environment.assetsPath}/prompt/raycast_llama2_7b`],
-    ["raycast_llama2:13b", `${environment.assetsPath}/prompt/raycast_llama2_13b`],
-  ]);
-  const body: OllamaApiCreateRequestBody = {
+export async function OllamaApiDelete(model: string): Promise<void> {
+  const url = "http://127.0.0.1:11434/api/delete";
+  const body = {
     name: model,
-    path: ModelFileMap.get(model) as string,
   };
 
-  return fetch(url, {
-    method: "POST",
+  await fetch(url, {
+    method: "DELETE",
+    headers: {
+      "Content-Type": "application/json",
+    },
     body: JSON.stringify(body),
   })
     .then((response) => {
-      if (response.ok) {
-        return response.text();
+      if (!response.ok) {
+        throw new ErrorOllamaModelNotInstalled(MessageOllamaModelNotInstalled.message, model);
       }
-
-      throw new ErrorOllamaCustomModel(ErrorOllamaCantCreateCustomModel.message, body.name, body.path);
     })
-    .then(() => true)
     .catch((err) => {
-      if (err instanceof ErrorOllamaCustomModel) {
+      if (err instanceof ErrorOllamaModelNotInstalled) {
         throw err;
       }
       console.error(err);
@@ -87,17 +67,67 @@ function OllamaApiCreateModel(model: string): Promise<boolean> {
 }
 
 /**
+ * Pull model.
+ * @param {string} model Model name.
+ * @returns {Promise<EventEmitter>} Response from the Ollama API with an EventEmitter with three event:
+ *  - {string} `message` - Message from the Ollama API.
+ *  - {string} `error` - Error from the Ollama API.
+ *  - {number} `downloading` - Downloading percentage.
+ *  - {string} `done` - On download completed.
+ */
+export async function OllamaApiPull(model: string): Promise<EventEmitter> {
+  const url = "http://127.0.0.1:11434/api/pull";
+  const body = {
+    name: model,
+  };
+  let emitter: EventEmitter | undefined;
+
+  while (emitter === undefined) {
+    emitter = await fetch(url, {
+      method: "POST",
+      body: JSON.stringify(body),
+    })
+      .then(async (response) => response.body)
+      .then((body) => {
+        if (body === undefined) {
+          return undefined;
+        }
+
+        const e = new EventEmitter();
+
+        body?.on("data", (chunk) => {
+          if (chunk !== undefined) {
+            const buffer = Buffer.from(chunk);
+            const json: OllamaApiPullResponse = JSON.parse(buffer.toString());
+            if (json.total && json.completed) {
+              e.emit("downloading", json.completed / json.total);
+            } else if (json.status === "success") {
+              e.emit("done", "Download completed");
+            } else if (json.error) {
+              e.emit("error", json.error);
+            } else {
+              e.emit("message", json.status);
+            }
+          }
+        });
+
+        return e;
+      })
+      .catch((err) => {
+        console.error(err);
+        throw ErrorOllamaNotInstalledOrRunning;
+      });
+  }
+  return emitter;
+}
+
+/**
  * Perform text generation with the selected model.
- * @param {string} prompt - Prompt to be used in the generation.
- * @param {string} model - Model name. Need to be installed or one of the following: raycast_orca:3b, raycast_llama2:7b, raycast_llama2:13b.
+ * @param {OllamaApiGenerateRequestBody} body - Ollama Generate Body Request.
  * @returns {Promise<EventEmitter>} Response from the Ollama API with an EventEmitter with two event: `data` where all generated text is passed on `string` format and `done` when inference is finished returning a `OllamaApiGenerateResponseDone` object contains all metadata of inference.
  */
-export async function OllamaApiGenerate(prompt: string, model: string): Promise<EventEmitter> {
+export async function OllamaApiGenerate(body: OllamaApiGenerateRequestBody): Promise<EventEmitter> {
   const url = "http://127.0.0.1:11434/api/generate";
-  const body: OllamaApiGenerateRequestBody = {
-    model: model,
-    prompt: prompt,
-  };
   let emitter: EventEmitter | undefined;
 
   while (emitter === undefined) {
@@ -111,19 +141,7 @@ export async function OllamaApiGenerate(prompt: string, model: string): Promise<
         }
 
         if (response.status === 400) {
-          return await OllamaApiTagsVerify(model).then(async (modelIsInstalled) => {
-            if (modelIsInstalled[0]) {
-              await OllamaApiCreateModel(model)
-                .then(() => {
-                  return undefined;
-                })
-                .catch((err) => {
-                  throw err;
-                });
-            } else {
-              throw new ErrorOllamaModelNotInstalled(MessageOllamaModelNotInstalled.message, modelIsInstalled[1]);
-            }
-          });
+          throw new ErrorOllamaModelNotInstalled(MessageOllamaModelNotInstalled.message, body.model);
         }
       })
       .then((body) => {
@@ -158,4 +176,51 @@ export async function OllamaApiGenerate(prompt: string, model: string): Promise<
       });
   }
   return emitter;
+}
+
+/**
+ * Perform text embeddings with the selected model.
+ * @param {string} prompt - text used for embeddings.
+ * @param {string} model - Model name.
+ * @returns {Promise<OllamaApiEmbeddingsResponse>} Ollama API embeddings response.
+ */
+export async function OllamaApiEmbeddings(prompt: string, model: string): Promise<OllamaApiEmbeddingsResponse> {
+  const url = "http://127.0.0.1:11434/api/embeddings";
+  const body: OllamaApiGenerateRequestBody = {
+    model: model,
+    prompt: prompt,
+  };
+  let embeddings: OllamaApiEmbeddingsResponse | undefined;
+
+  while (embeddings === undefined) {
+    embeddings = await fetch(url, {
+      method: "POST",
+      body: JSON.stringify(body),
+    })
+      .then(async (response) => {
+        if (response.ok) {
+          return response;
+        }
+
+        if (response.status === 400) {
+          throw new ErrorOllamaModelNotInstalled(MessageOllamaModelNotInstalled.message, body.model);
+        }
+      })
+      .then(async (response) => {
+        if (response === undefined) {
+          return undefined;
+        }
+        return (await response.json()) as OllamaApiEmbeddingsResponse;
+      })
+      .catch((err) => {
+        if (err instanceof ErrorOllamaModelNotInstalled) {
+          throw err;
+        } else if (err instanceof ErrorOllamaCustomModel) {
+          throw err;
+        }
+        console.error(err);
+        throw ErrorOllamaNotInstalledOrRunning;
+      });
+  }
+  return embeddings;
 }