+++ extensions/raycast-ollama/src/ollama-custom-ask.tsx
@@ -1,12 +1,11 @@
-import { RaycastArgumentsOllamaAskCustom } from "./api/types";
-import { GetCustomPrompt } from "./api/prompt";
-import ResultView from "./api/main";
-import { getPreferenceValues } from "@raycast/api";
-
-const preferences = getPreferenceValues();
+import { OllamaApiGenerateRequestBody, RaycastArgumentsOllamaAskCustom } from "./api/types";
+import { ResultView } from "./api/main";
 
 export default function Command(props: RaycastArgumentsOllamaAskCustom): JSX.Element {
-  const prompt = GetCustomPrompt(preferences.ollamaAskCustomPromptModel, props.arguments.prompt, props.arguments.query);
+  const body = {
+    model: props.arguments.model,
+    prompt: props.arguments.query,
+  } as OllamaApiGenerateRequestBody;
 
-  return ResultView(preferences.ollamaAskCustomPromptModel, prompt.prompt, prompt.tagEnd, false);
+  return ResultView(body, false);
 }