+++ extensions/tesla/src/turnSentryOn.tsx
@@ -1,0 +1,5 @@
+import ViewCar from "./viewCar";
+
+export default function Command() {
+  return <ViewCar command="enableSentryMode" loadingMessage="Enabling Sentry Mode..." />;
+}