+++ extensions/whois/raycast-env.d.ts
@@ -24,3 +24,13 @@ declare namespace Arguments {
   "domain": string
 }
 }
+
+
+declare module "swift:*" {
+  function run<T = unknown, U = any>(command: string, input?: U): Promise<T>;
+  export default run;
+	export class SwiftError extends Error {
+    stderr: string;
+    stdout: string;
+  }
+}