+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,9 @@
 # reclaim Changelog
 
+## [New Command] - 2023-08-10
+
+- Added new `Search Tasks` command
+
 ## [Update] - 2023-08-02
 
 - Added filter to remove declined events in MenuBar.